describe the fork toward the top of README
This is a fork for testing and merging changes I hope to push back upstream.  While it's likely to be a little more long-lived then a typical PR fork, it's not going to be around too long; I'll drop it ones I have changes merged upstream (or give up on that).